fix broken spec, replace ballerina with dancer, fix spelling of MB name dance_spec was passing in dance instance in test for extended class methods. change to pass in Dance class
