+++ bin/dance_party
@@ -1,8 +1,8 @@
 require_relative "../lib/kid.rb"
-require_relative "../lib/ballerina.rb"
+require_relative "../lib/dancer.rb"
 
 angelina = Kid.new("Anglina")
-mikhail_barishnkov = Ballerina.new("Mikhail")
+mikhail_baryshnikov = Dancer.new("Mikhail")
 
 puts "#{angelina.name} says: #{angelina.twirl}"
-puts "#{mikhail_barishnkov.name} says: #{mikhail_barishnkov.take_a_bow}"
+puts "#{mikhail_baryshnikov.name} says: #{mikhail_baryshnikov.take_a_bow}"
