+++ bin/extending
@@ -1,5 +1,5 @@
 require_relative "../lib/kid.rb"
-require_relative "../lib/ballerina.rb"
+require_relative "../lib/dancer.rb"
 
 puts Kid.metadata
-puts Ballerina.metadata
+puts Dancer.metadata
