--- lib/ballerina.rb
@@ -1,12 +1,0 @@
-require_relative './class_methods_module.rb'
-require_relative './dance_module.rb'
-
-class Ballerina
-  include Dance
-  extend MetaDancing
-  attr_accessor :name
-
-  def initialize(name)
-    @name = name
-  end
-end
