+++ lib/class_methods_module.rb
@@ -1,5 +1,5 @@
 module MetaDancing
   def metadata
-    "This class produces object that love to dance."
+    "This class produces objects that love to dance."
   end
 end
