+++ lib/dancer.rb
@@ -1,0 +1,15 @@
+require_relative './dance_module.rb'
+require_relative './class_methods_module.rb'
+require_relative './fancy_dance.rb'
+
+class Dancer
+  # include Dance
+  # extend MetaDancing
+  include FancyDance::InstanceMethods
+  extend FancyDance::ClassMethods
+  attr_accessor :name
+
+  def initialize(name)
+    @name = name
+  end
+end
