+++ spec/spec_helper.rb
@@ -1,8 +1,9 @@
-require_relative '../lib/ballerina'
 require_relative '../lib/class_methods_module'
+require_relative '../lib/dancer'
 require_relative '../lib/dance_module'
 require_relative '../lib/kid'
 require_relative '../lib/fancy_dance.rb'
+require 'pry'
 
 RSpec::Matchers.define :have_instance_dance_methods do
   match do |actual|
